requirements.txt
asgiref==3.3.4
Django==3.2.2
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.4
djangorestframework-csv==2.1.0
gunicorn==20.1.0
psycopg2==2.8.6
six==1.16.0
unicodecsv==0.14.1
